refactor: move to more explicit testdirectory
