--- processor/src/test/java/org/mapstruct/ap/test/bugs/_1140/EmptyJar.java
@@ -1,4 +1,0 @@
-package org.mapstruct.ap.test.bugs._1140;
-
-public class EmptyJar {
-}
