+++ processor/src/test/java/org/mapstruct/ap/test/emptytarget/AirplaneWithNoAccessors.java
@@ -1,4 +1,4 @@
-package org.mapstruct.ap.test.bugs._1140;
+package org.mapstruct.ap.test.emptytarget;
 
 public class AirplaneWithNoAccessors {
     private int flightNumber;
