+++ processor/src/test/java/org/mapstruct/ap/test/emptytarget/EmptyJar.java
@@ -1,0 +1,4 @@
+package org.mapstruct.ap.test.emptytarget;
+
+public class EmptyJar {
+}
